[Maps] Reorder page elements in print view
